Move statements out of Client
This reverts e071532d350c332e48d1fbe8a6a81712e789f574
